+++ core/src/main/scala/com/raphtory/internals/graph/GraphPartition.scala
@@ -21,7 +21,7 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
 
   // Ingesting Vertices
   def addVertex(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
@@ -29,10 +29,10 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
       vertexType: Option[Type]
   ): Unit
 
-  def removeVertex(sourceID: Int, msgTime: Long, index: Long, srcId: Long): List[GraphUpdateEffect]
+  def removeVertex(sourceID: Long, msgTime: Long, index: Long, srcId: Long): List[GraphUpdateEffect]
 
   def inboundEdgeRemovalViaVertex(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
@@ -40,7 +40,7 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
   ): GraphUpdateEffect
 
   def outboundEdgeRemovalViaVertex(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
@@ -49,7 +49,7 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
 
   // Ingesting Edges
   def addEdge(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
@@ -59,7 +59,7 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
   ): Option[GraphUpdateEffect]
 
   def syncNewEdgeAdd(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
@@ -70,7 +70,7 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
   ): GraphUpdateEffect
 
   def syncExistingEdgeAdd(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
@@ -79,7 +79,7 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
   ): GraphUpdateEffect
 
   def batchAddRemoteEdge(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
@@ -88,17 +88,17 @@ abstract private[raphtory] class GraphPartition(graphID: String, partitionID: In
       edgeType: Option[Type]
   ): Unit
 
-  def removeEdge(sourceID: Int, msgTime: Long, index: Long, srcId: Long, dstId: Long): Option[GraphUpdateEffect]
+  def removeEdge(sourceID: Long, msgTime: Long, index: Long, srcId: Long, dstId: Long): Option[GraphUpdateEffect]
 
   def syncNewEdgeRemoval(
-      sourceID: Int,
+      sourceID: Long,
       msgTime: Long,
       index: Long,
       srcId: Long,
       dstId: Long,
       srcRemovals: List[(Long, Long)]
   ): GraphUpdateEffect
-  def syncExistingEdgeRemoval(sourceID: Int, msgTime: Long, index: Long, srcId: Long, dstId: Long): GraphUpdateEffect
+  def syncExistingEdgeRemoval(sourceID: Long, msgTime: Long, index: Long, srcId: Long, dstId: Long): GraphUpdateEffect
 
   def syncExistingRemovals(
       msgTime: Long,
